# api/openapi.yaml
openapi: 3.1.0
info:
  title: ShowMesh Control API
  version: "1"
  summary: >-
    The coordinator's public, versioned control API and Server-Sent
    Events change stream (ADR-014).
  description: >-
    This is the versioned public contract ADR-014 requires: it is designed
    to be used without any Operator UI, by a CLI (`showmeshctl`), an
    automation system, or an alternate client. Everything here is served
    under `/api/v1`; the container healthcheck endpoints `/healthz`,
    `/readyz`, and `/version` are infrastructure probes outside this
    versioned contract and are not documented here.


    **Versioning is additive-only within v1.** Fields may be added to any
    response; they are never removed, renamed, or retyped. A client MUST
    ignore fields it does not recognize. A breaking change is a `/api/v2`
    contract, not an edit to this one.


    **`serverTime` appears on every response, including every error.** A
    client computes evidence ages against `serverTime` rather than its own
    clock, so clock skew between the client and the coordinator is visible
    instead of silently wrong. Payloads never carry a precomputed age or
    "secondsAgo" field for exactly this reason: an age field would change on
    every poll even when nothing else did, which would make change
    detection on the SSE stream fire continuously for no real change.


    **`offline` means the control-plane (MQTT) connection is gone, not that
...


...
    `state: "unknown_age"`.


...


...


...


...

...
servers:
...
security:
  - bearerAuth: []
  - cookieAuth: []
  - {}
paths:
...
  /fpp/{instanceId}/commands:
    post:
      operationId: dispatchFPPCommand
      summary: Dispatch a primitive FPP lifecycle command (Step 7 seam C, Step 8)
      description: >-
        Behind `fpp:command` (ADR-024 decision 4). Dispatches FPP's own
        native command over POST at its own /api/command endpoint
        (ADR-001: this is ShowMesh invoking FPP's own command, never a
        second scheduler) and confirms by evidence against the collector's
        observations before answering, up to an internal, per-action
        deadline (ADR-003). A `200` from FPP is never reported as this
        endpoint's own success on its own — see FPPCommandResult.outcome. A
        cookie-authenticated request additionally requires
        `Sec-Fetch-Site: same-origin` (ADR-024 decision 6, `403` on
        absence, same Problem schema as a missing-scope `403`); a
        bearer-token-authenticated request is exempt.

        `idempotencyKey` is resolved FIRST, before `startPlaylist`'s own
        `ifBusy` guard (or any other primitive's pre-dispatch guard) ever
        runs — a review finding proved live that the reverse order answers
        a legitimate replay with a fresh `409` the guard invented for what
        it wrongly evaluated as a brand-new request. A replayed
        `idempotencyKey` — reused against the SAME `action`, the SAME
        `instanceId`, and the SAME normalized `params` it was first used
        against — dispatches nothing and returns the original command's
        result, flagged `replay: true`, with no guard evaluated at all. An
        `idempotencyKey` reused against a DIFFERENT `action`, a DIFFERENT
        `instanceId`, or DIFFERENT `params` is a `409` conflict, not a
        replay, also decided before any guard runs: see the
        `idempotencyKey` field on any FPPCommandRequest variant below (the
        text is identical on all eight) and the `409` response below. Only
        a genuinely NEW `idempotencyKey` — one this coordinator has never
        recorded — reaches a primitive's own guard at all.

        `startPlaylist`'s own `ifBusy` default (`"refuse"`) additionally
        produces a `409` BEFORE anything is dispatched to FPP when a
        DIFFERENT playlist is currently confirmed playing (`type`
        `fpp-start-playlist-busy`, `detail` names what is playing), or when
        the evidence needed to decide that is not current (`type`
        `fpp-start-playlist-evidence-not-current` — this coordinator never
        proceeds on the grounds that it could not tell) — two DIFFERENT
        `type` values on the same `409` status, neither of them the plain
        `conflict` the idempotency-key case above uses, so a client can
        tell all three apart without parsing `detail` prose; see
        StartPlaylistCommandRequest.params's own description. This is a
        guard against an accidental interruption, not a lock: it is
        evaluated against evidence that can be stale, and cannot prevent a
        race against FPP's own scheduler.
      parameters:
        - name: instanceId
          in: path
          required: true
          schema:
            type: string
          description: Same ID syntax as a node ID (contract section 7).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FPPCommandRequest"
      responses:
        "200":
          description: OK
          headers:
            ShowMesh-API-Version:
              $ref: "#/components/headers/ShowMesh-API-Version"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FPPCommandResponse"
        "400":
          $ref: "#/components/responses/InvalidParameter"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "409":
          description: >-
            Three DISTINCT causes, three DISTINCT `type` values, decided in
            this order (see this operation's own description): (1)
            `idempotencyKey` was already used for a command whose `action`,
            `instanceId`, or normalized `params` differs from this
            request's own — a conflict, not a replay, `type` `conflict`,
            remedy "mint a fresh key"; decided BEFORE either guard below
            ever runs. Only once the key is confirmed genuinely new does
            `startPlaylist`'s own `ifBusy` guard get to refuse: (2) a
            DIFFERENT playlist is confirmed playing, `type`
            `fpp-start-playlist-busy`, remedy "resend with
            ifBusy: replace"; or (3) the evidence needed to decide that is
            not current, `type` `fpp-start-playlist-evidence-not-current`,
            remedy "retry once evidence is current, or resend with
            ifBusy: replace if interrupting is intended". No two of these
            three share a `type`, specifically so a client branches on
            `type`, never on `detail` prose, since (1)'s remedy is the
            OPPOSITE of (2)'s and (3)'s.
          headers:
            ShowMesh-API-Version:
              $ref: "#/components/headers/ShowMesh-API-Version"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          description: >-
            ADR-024 decision 11's fail-closed default: this action is not
            a member of decision 11's blackout/stop/power-off safety
            class, and the pre-dispatch write that must durably record a
            command before it is dispatched could not be appended to this
            coordinator's audit store. `type`
            `fpp-command-refused-audit-unavailable`. Nothing was recorded
            and nothing was dispatched to FPP; retry once the audit store
            is writable again.
          headers:
            ShowMesh-API-Version:
              $ref: "#/components/headers/ShowMesh-API-Version"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  responses:
...
    ResourceNotFound:
      description: No resource with the given identifier exists.
      headers:
        ShowMesh-API-Version:
          $ref: "#/components/headers/ShowMesh-API-Version"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    InvalidParameter:
      description: A path or query parameter failed validation.
      headers:
        ShowMesh-API-Version:
          $ref: "#/components/headers/ShowMesh-API-Version"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: >-
        No valid credential was presented: reads are closed and none was
        given (or it did not authenticate), or the route is a write /
        `GET /audit` and no credential authenticated at all.
      headers:
        ShowMesh-API-Version:
          $ref: "#/components/headers/ShowMesh-API-Version"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: >-
        A valid credential authenticated, but the principal does not hold
        the scope this operation requires. `detail` names the missing
        scope (ADR-024 decision 4).
      headers:
        ShowMesh-API-Version:
          $ref: "#/components/headers/ShowMesh-API-Version"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    MethodNotAllowed:
      description: >-
        The path names a real route, but this request's method is not one
        that route serves. A method this document does not define for a
        path returns this response rather than a `404`, and the `Allow`
        header names exactly the methods that path does serve.

        This description used to say that the document defined only `get`
        operations because every route was read-only. That stopped being
        true in Step 6, which added `POST /session` and
        `POST /bootstrap`, and stopped being true three times over in
        Step 7, which added the configuration, discovery and FPP command
        writes. What has NOT changed is the rule underneath it: no state
        change in this API is reachable by `GET` (ADR-024), so a `GET`
        remains safe to issue against any path here.
      headers:
        ShowMesh-API-Version:
          $ref: "#/components/headers/ShowMesh-API-Version"
        Allow:
          $ref: "#/components/headers/Allow"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    InternalError:
      description: >-
        The coordinator failed to answer this request for a reason that
        has nothing to do with whether the requested resource exists or
        the request was well formed — an internal dependency (e.g. the
        store) failed. `detail` never carries the underlying error; it is
        logged coordinator-side only.
      headers:
        ShowMesh-API-Version:
          $ref: "#/components/headers/ShowMesh-API-Version"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  headers:
    ShowMesh-API-Version:
      description: >-
        The API major version this response was served by. Present on
        every `/api/v1` response, success or error, with no exception
        (contract section 6.2). A client MAY also send this as a request
        header naming the version it expects; a coordinator that does not
        serve that version answers with the `UnsupportedAPIVersion`
        problem instead of a partial or best-guess render (contract
        section 6.6).
      schema:
        type: string
        enum: ["1"]
    Allow:
      description: >-
        Present only on a `405` response. The exact, comma-separated set
        of HTTP methods the requested path serves (e.g. `"GET, HEAD"`),
        computed from the routes actually registered — never a static or
        hand-maintained list.
      schema:
        type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        `Authorization: Bearer <token>` (ADR-024). token is an API token
        minted for a principal by an admin, displayed exactly once at
        creation and stored coordinator-side only as a SHA-256 digest. A
        malformed or non-`Bearer` `Authorization` header never falls
        through to `cookieAuth` (ADR-024 decision 6) — it is treated as no
        credential at all. Reads are open by default (this scheme is
        optional for them, per the top-level `security` requirement not
        being exclusive); every write requires it or `cookieAuth`, with no
        opt-out. `/healthz`, `/readyz`, and `/version` are outside this
        contract entirely and are never authenticated, at any
        configuration.
    cookieAuth:
      type: apiKey
      in: cookie
      name: showmesh_session
      description: >-
        The `HttpOnly` session cookie `POST /session` mints (ADR-024
        decision 5). Not usable outside a browser in practice (nothing
        exposes it to script), and a write authenticated this way
        additionally requires the `Sec-Fetch-Site: same-origin` request
        header (decision 6) — omitted from this scheme's own contract
        because OpenAPI's securityScheme vocabulary has no way to express
        a conditional required header; see this document's top-level
        description and `components.responses.CSRFRejected`.
  schemas:
...
    FPPCommandRequest:
      description: >-
        The body of POST /fpp/{instanceId}/commands (Step 7 seam C, Step 8,
        ADR-001, ADR-003) — a discriminated union on `action`, one member
        per docs/bench/fpp-command-vocabulary.md section 4's eight
        primitives. This is a deliberate correction: earlier this schema
        declared `params` as a bare, propertyless `object`, which a
        strict-JSON-Schema code generator renders as a type NO non-empty
        object satisfies — the exact opposite of what most of these eight
        actions need, since three of them (`startPlaylist`,
        `stopPlaylistGracefully`, `setVolume`) take real parameters. Each
        variant below carries its own concrete `params` shape — required
        fields, enums, numeric bounds — as real JSON Schema, not as prose
        a generator cannot see.

        `idempotencyKey` is required on every variant (ARCHITECTURE
        section 8.1): RES-015 section 7.3 established that FPP supplies
        nothing a coordinator-minted key could be derived from, so the
        caller (showmeshctl, the Operator UI) mints a fresh one per
        invocation. Its scoping rule (replay vs. `409` conflict) is
        identical on every variant — see any one variant's own
        `idempotencyKey` description below; the text is not repeated here
        a ninth time.

        `params` is REQUIRED whenever an action has at least one required
        parameter (`startPlaylist`, `setVolume`) and OPTIONAL otherwise.
        When optional, absent or `{}` means every parameter takes its
        documented default. An explicit `"params": null` is ALWAYS a
        `400`, for every action — every variant's own `params` schema is a
        plain JSON object type, never nullable, so a generated client
        cannot even construct this shape, and this coordinator's own
        handler rejects it explicitly rather than treating null as
        "omitted" (an explicit null is not the same as an omitted field).
        A required parameter absent, present as `null`, or (for a string)
        present as `""` are three DIFFERENT `400` responses, each naming
        exactly what is wrong. A key an action's own `params` schema does
        not declare is a `400` naming it (every variant's `params` sets
        `additionalProperties: false`) — deliberately stricter than this
        document's own "clients ignore unknown fields" rule, which governs
        a client reading RESPONSES, not a server accepting a WRITE: a
        typo'd parameter name must never silently fall back to that
        parameter's own default.
      discriminator:
        propertyName: action
        mapping:
          startPlaylist: "#/components/schemas/StartPlaylistCommandRequest"
          stopPlaylistGracefully: "#/components/schemas/StopPlaylistGracefullyCommandRequest"
          setVolume: "#/components/schemas/SetVolumeCommandRequest"
          stopPlaylist: "#/components/schemas/NoParamsFPPCommandRequest"
          pausePlaylist: "#/components/schemas/NoParamsFPPCommandRequest"
          resumePlaylist: "#/components/schemas/NoParamsFPPCommandRequest"
          nextPlaylistItem: "#/components/schemas/NoParamsFPPCommandRequest"
          prevPlaylistItem: "#/components/schemas/NoParamsFPPCommandRequest"
      oneOf:
        - $ref: "#/components/schemas/StartPlaylistCommandRequest"
        - $ref: "#/components/schemas/StopPlaylistGracefullyCommandRequest"
        - $ref: "#/components/schemas/SetVolumeCommandRequest"
        - $ref: "#/components/schemas/NoParamsFPPCommandRequest"
...
    StopPlaylistGracefullyCommandRequest:
      type: object
      description: >-
        `action: "stopPlaylistGracefully"` (docs/bench/fpp-command-vocabulary.md
        section 3.3/4): a graceful stop's terminal state is bounded by the
        currently playing item's own remaining runtime, not by any number
        this coordinator can choose — see FPPCommandResult.outcomeReason.
      required: [action, idempotencyKey]
      properties:
        action:
          type: string
          const: stopPlaylistGracefully
        idempotencyKey:
          type: string
          description: >-
            Scoped to the exact (action, instanceId, normalized params)
            triple it is first used against — a database-level UNIQUE
            constraint on this value alone (schemaV6) cannot express that
            scope, so the server checks it explicitly: reusing a key
            against the SAME action, instanceId, AND normalized params is
            a replay (nothing is dispatched a second time, the original
            result is returned, flagged `replay: true`); reusing it
            against a DIFFERENT action, a DIFFERENT instanceId, or
            DIFFERENT params is a `409` conflict, refused outright, never
            answered as if it belonged to whichever command first claimed
            the key. Mint a fresh key per genuinely new request — never
            reuse one across two different invocations "to be safe."
        params:
          type: object
          description: >-
            Optional; absent or `{}` means `afterLoop` defaults to
            `false`.
          additionalProperties: false
          properties:
            afterLoop:
              type: boolean
              default: false
    SetVolumeCommandRequest:
      type: object
      description: >-
        `action: "setVolume"` (docs/bench/fpp-command-vocabulary.md
        section 3.6/4).
      required: [action, idempotencyKey, params]
      properties:
        action:
          type: string
          const: setVolume
        idempotencyKey:
          type: string
          description: >-
            Scoped to the exact (action, instanceId, normalized params)
            triple it is first used against — a database-level UNIQUE
            constraint on this value alone (schemaV6) cannot express that
            scope, so the server checks it explicitly: reusing a key
            against the SAME action, instanceId, AND normalized params is
            a replay (nothing is dispatched a second time, the original
            result is returned, flagged `replay: true`); reusing it
            against a DIFFERENT action, a DIFFERENT instanceId, or
            DIFFERENT params is a `409` conflict, refused outright, never
            answered as if it belonged to whichever command first claimed
            the key. Mint a fresh key per genuinely new request — never
            reuse one across two different invocations "to be safe."
        params:
          type: object
          required: [volume]
          additionalProperties: false
          properties:
            volume:
              type: integer
              minimum: 0
              maximum: 100
              description: >-
                A JSON number with a fractional part is a `400`, never
                truncated (capture section 1.5: FPP itself silently
                coerces and clamps a bad value; this coordinator does not
                repeat that for its own parameters).
...
    FPPCommandResponse:
      type: object
      description: The body of a successful (200) response from POST /fpp/{instanceId}/commands.
      required: [serverTime, command]
      properties:
        serverTime:
          type: string
          format: date-time
        command:
          $ref: "#/components/schemas/FPPCommandResult"
...